Updated project template
