+++ src/Templates/PerspexApplicationTemplate/App.paml
@@ -1,0 +1,5 @@
+﻿<Application xmlns="https://github.com/perspex">
+  <Application.Styles>
+
+  </Application.Styles>
+</Application>
