+++ src/Templates/PerspexApplicationTemplate/App.paml.cs
@@ -1,9 +1,9 @@
 ﻿using System;
-
 using Perspex;
 using Perspex.Controls;
 using Perspex.Diagnostics;
 using Perspex.Themes.Default;
+using Perspex.Markup.Xaml;
 
 namespace $safeprojectname$
 {
@@ -14,6 +14,15 @@ namespace $safeprojectname$
             RegisterServices();
             InitializeSubsystems((int)Environment.OSVersion.Platform);
             Styles = new DefaultTheme();
+            PerspexXamlLoader.Load(this);
+        }
+
+        static void Main(string[] args)
+        {
+            var app = new App();
+            var window = new MainWindow();
+            window.Show();
+            app.Run(window);
         }
 
         public static void AttachDevTools(Window window)
