--- src/Templates/PerspexApplicationTemplate/Program.cs
@@ -1,19 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-
-namespace $safeprojectname$
-{
-    class Program
-    {
-        static void Main(string[] args)
-        {
-            var app = new App();
-            var window = new MainWindow();
-            window.Show();
-            app.Run(window);
-        }
-    }
-}
